Methods to manage permissions

+++ src/Application/Users/Commands/Delete/DeleteUser.cs
@@ -13,6 +13,7 @@
 //  limitations under the License.
 //
 
+using Common.Application.Interfaces;
 using TrackHub.Security.Application.Users.Events;
 
 namespace TrackHub.Security.Application.Users.Commands.Delete;
@@ -20,12 +21,18 @@ namespace TrackHub.Security.Application.Users.Commands.Delete;
 [Authorize(Resource = Resources.Users, Action = Actions.Delete)]
 public readonly record struct DeleteUserCommand(Guid Id) : IRequest;
 
-public class DeleteUserCommandHandler(IUserWriter writer, IPublisher publisher) : IRequestHandler<DeleteUserCommand>
+public class DeleteUserCommandHandler(IUserWriter writer, IUser user, IPublisher publisher) : IRequestHandler<DeleteUserCommand>
 {
+    private Guid UserId { get; } = user.Id is null ? throw new UnauthorizedAccessException() : new Guid(user.Id);
 
     // Handle method to delete a user
+    // The user cannot delete itself
     public async Task Handle(DeleteUserCommand request, CancellationToken cancellationToken)
     {
+        if (request.Id == UserId)
+        {
+            throw new UnauthorizedAccessException("You cannot delete yourself");
+        }
         await writer.DeleteUserAsync(request.Id, cancellationToken);
         await publisher.Publish(new UserDeleted.Notification(request.Id), cancellationToken);
     }
